% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_donorpool.R
\name{draw_donorpool}
\alias{draw_donorpool}
\title{Draw Optimaized Counterparts from the Donor Pool.}
\usage{
draw_donorpool(target, donorpool, criteria = "BIC", nvmax)
}
\arguments{
\item{target}{The unit we want to match in the pre-treatment and to predict in the after-treatment.}

\item{donorpool}{The candidates of donorpool we want to optimizaed.}

\item{criteria}{The information criteria we want to use.}

\item{nvmax}{The maxmum number we want to draw from the donor pool}
}
\value{
The list of selected optimale counterparts and the $R^2$ of the model
}
\description{
\code{draw_donorpool} returns the best couterparts that match the target unit in the pre-treatment period.
}
