% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hcw_data}
\alias{hcw_data}
\title{The Real GDP Growth Rate of Different Countries, from 1993Q1 to 2008Q1.}
\format{A data frame with 61 rows and 26 variables:
\describe{
  \item{date}{Year-Quarter Date Variable}
  ...
}}
\source{
Cheng Hsiao, H. Steve Ching, and Shui Ki Wan (2012), "A panel data approach for program evaluation: measuring the benefits of political and economic integration of Hong kong with mainland China", Journal of Applied Econometrics.
}
\usage{
hcw_data
}
\description{
A dataset containing the Real GDP Growth Rate of 25 countries or regions.
}
\keyword{datasets}
