% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo_time.R
\name{placebo_time}
\alias{placebo_time}
\title{Placebo Test Using Fake Treatment Time}
\usage{
placebo_time(result, lead_period)
}
\arguments{
\item{result}{The object generated by the \code{\link{draw_donorpool}}.}

\item{lead_period}{how many time points in advance of the real treatment time we want to test.
For example, if the real treatment time is t, and we want to test t - 1, t - 2, t - 3,
we could set the lead_period = 3}
}
\value{
The ggplot object visualizing the placebo test using fake treatment time
}
\description{
\code{placebo_time} is the function of doing placebo test by giving the fake treatment time to see
 if the treatment effect we got using real treatment time is the largest, in other words, we are expecting
 to see the fake treatemt time will induce smaller delta.
}
