% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo_unit.R
\name{placebo_unit}
\alias{placebo_unit}
\title{Placebo Test Using Fake Treatment Unit}
\usage{
placebo_unit(result)
}
\arguments{
\item{result}{The object generated by the \code{\link{draw_donorpool}}.}
}
\value{
The ggplot object visualizing the placebo test using fake treatment units
}
\description{
\code{placebo_unit} is the function of doing placebo test by using fake treatment units to see
 if the treatment effect we got using real treatment unit is the largest, in other words, we are expecting
 to see the fake treatemt unit will induce smaller delta.
}
