% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_delta.R
\name{plot_delta}
\alias{plot_delta}
\title{Visualization of the difference between the simulated counterfactual and actural value.}
\usage{
plot_delta(result)
}
\arguments{
\item{result}{The object generated by the \code{\link{draw_donorpool}}.}
}
\value{
The ggplot object visualizing the delta.
}
\description{
\code{plot_delta} is the visualization of the difference between the simulated counterfactual
(generated by \code{\link{draw_donorpool}}) and actural value.
}
