% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_result.R
\name{plot_sim_result}
\alias{plot_sim_result}
\title{Visualization of the simulation results.}
\usage{
plot_sim_result(result)
}
\arguments{
\item{result}{The object generated by the \code{\link{draw_donorpool}}.}
}
\value{
The ggplot object visualizing the simulation results.
}
\description{
\code{plot_sim_result} is the visualization of the results generated by \code{\link{draw_donorpool}}.
It contains both the actural and conunterfactual value of target across treatment time point.
}
