% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_donorpool.R
\name{draw_donorpool}
\alias{draw_donorpool}
\title{Draw Optimaized Counterparts from the Donor Pool.}
\usage{
draw_donorpool(data, target_name, donorpool_name = NULL, time_name, period,
  criteria = "BIC", nvmax)
}
\arguments{
\item{data}{The imput data, which contains the outcome Y for both the target and donor pool units,
and also a time variable}

\item{target_name}{The unit name we want to match in the pre-treatment and to predict in the after-treatment.}

\item{donorpool_name}{The unit name of candidates in donorpool we want to optimizaed.}

\item{time_name}{The name of time variable.}

\item{period}{The number of observations in pre-treatment period}

\item{criteria}{The information criteria we want to use.}

\item{nvmax}{The maxmum number we want to draw from the donor pool}
}
\value{
The list of selected optimale counterparts and the $R^2$ of the model
}
\description{
\code{draw_donorpool} returns the best couterparts that match the target unit in the pre-treatment period.
}
